Fixed broken links to w3c-test.org (specs and test suites)
A number of specifications and test suites got moved to different folders on
w3c-test.org. This commit should fix all broken links that appear in the
final roadmap.

As an exception to the rule, I could not find the tests referenced by the
inputaccept.json file for the Media Capture specification. I see there's
an on-going pull request at:
  https://github.com/w3c/web-platform-tests/pull/306
... but files do not appear under http://w3c-test.org/submissions. The
broken link remains.

Not related but note the notifications API tutorial is no longer available
on WebPlatform.org so I dropped the section from the notification.json file.